Revert "stop fghting testground. For now"
This reverts commit a976c5371b84968d70a24610b5a71df529105ba5.
