+++ Cargo.toml
@@ -3,6 +3,9 @@
 members = [
     "smoke",
     "episub",
+    
+    "utils/gen_topology",
+    "utils/gen_topology_files"
 ]
 
 [workspace.dependencies]
