+++ episub/16instances.dot
@@ -1,0 +1,38 @@
+digraph {
+	0 -> 10;
+	0 -> 12;
+	0 -> 5;
+	0 -> 14;
+	1 -> 14;
+	2 -> 0;
+	3 -> 4;
+	3 -> 5;
+	4 -> 1;
+	4 -> 12;
+	4 -> 6;
+	4 -> 13;
+	5 -> 12;
+	5 -> 13;
+	6 -> 11;
+	6 -> 5;
+	6 -> 2;
+	6 -> 1;
+	7 -> 3;
+	7 -> 11;
+	8 -> 10;
+	8 -> 1;
+	8 -> 9;
+	9 -> 3;
+	9 -> 15;
+	10 -> 12;
+	10 -> 1;
+	11 -> 1;
+	11 -> 3;
+	12 -> 2;
+	12 -> 15;
+	13 -> 7;
+	14 -> 11;
+	14 -> 3;
+	14 -> 5;
+	15 -> 11;
+}
