+++ episub/16instances.json
@@ -1,0 +1,1 @@
+{"params":{"seed":40,"total_validators":30,"total_nodes_with_vals":16,"total_nodes":16,"connections_range":{"start":3,"end":7}},"validator_assignments":{"0":[],"1":[28],"2":[14],"3":[9],"4":[0,29],"5":[19,27],"6":[4,6,8],"7":[3,16],"8":[5,12,13,17,18,26],"9":[15],"10":[7],"11":[2],"12":[11],"13":[1,10,21,24,25],"14":[20,22],"15":[23]},"outbound_peers":{"0":[10,12,5,14],"1":[14],"2":[0],"3":[4,5],"4":[1,12,6,13],"5":[12,13],"6":[11,5,2,1],"7":[3,11],"8":[10,1,9],"9":[3,15],"10":[12,1],"11":[1,3],"12":[2,15],"13":[7],"14":[11,3,5],"15":[11]}}
