+++ episub/Cargo.toml
@@ -18,13 +18,12 @@ testground = "0.4.0"
 tokio = { version = "1.21.2", features = ["macros"] }
 tracing = "0.1.35"
 tracing-subscriber = { version = "0.3.14", features = ["env-filter"] }
+# This is ridiculous
+gen_topology = { git = "https://github.com/sigp/gossipsub-testground", branch = "episub" }
+# gen_topology = { path = "../utils/gen_topology" }
 futures = "0.3.24"
 npg = { git = "https://github.com/sigp/eth-npg", rev = "0005322440f3f65a8c1f1d0f267a47c19b3320a1" }
-rand = "0.8.5"
-rand_chacha = "0.3.1"
-clap = { version = "4.0.10", features = ["derive"] }
 
-# Required because testground can't access the patches from the workspace. These are required to use the slot_clock crate
 [patch]
 [patch.crates-io]
 types = { git = "https://github.com/divagant-martian/lighthouse", rev = "22fda6cae62c858bd3b1c057700406a7bb358c97" }
