+++ episub/small.dot
@@ -1,0 +1,26 @@
+digraph {
+	0 -> 2;
+	0 -> 14;
+	1 -> 10;
+	1 -> 7;
+	1 -> 15;
+	2 -> 14;
+	2 -> 9;
+	2 -> 4;
+	3 -> 12;
+	4 -> 1;
+	4 -> 11;
+	5 -> 8;
+	6 -> 3;
+	7 -> 14;
+	7 -> 15;
+	8 -> 11;
+	9 -> 8;
+	10 -> 12;
+	11 -> 3;
+	13 -> 5;
+	13 -> 9;
+	13 -> 7;
+	14 -> 11;
+	15 -> 6;
+}
