+++ episub/small.json
@@ -1,0 +1,1 @@
+{"params":{"seed":40,"total_validators":200,"total_nodes_with_vals":6,"total_nodes":16,"connections_range":{"start":2,"end":4}},"validator_assignments":{"0":[16,18,26,34,37,38,51,64,67,69,70,78,99,107,108,114,130,163,180,186,187,194],"1":[5,7,24,27,29,31,54,60,66,74,77,81,82,85,96,97,112,115,120,123,125,139,150,151,152,154,156,161,164,165,167,195],"2":[12,22,25,35,44,50,76,103,126,131,138,190,196,197],"3":[14,28,36,40,56,72,100,135,189,192],"4":[4,19,23,32,33,46,62,63,79,87,89,91,92,93,94,119,124,133,134,136,142,147,148,155,157,162,169,173,193,198],"5":[0,1,2,3,6,8,9,10,11,13,15,17,20,21,30,39,41,42,43,45,47,48,49,52,53,55,57,58,59,61,65,68,71,73,75,80,83,84,86,88,90,95,98,101,102,104,105,106,109,110,111,113,116,117,118,121,122,127,128,129,132,137,140,141,143,144,145,146,149,153,158,159,160,166,168,170,171,172,174,175,176,177,178,179,181,182,183,184,185,188,191,199]},"outbound_peers":{"0":[2,14],"1":[10,7,15],"2":[14,9,4],"3":[12],"4":[1,11],"5":[8],"6":[3],"7":[14,15],"8":[11],"9":[8],"10":[12],"11":[3],"13":[5,9,7],"14":[11],"15":[6]}}
