+++ episub/src/main.rs
@@ -1,7 +1,3 @@
-#[allow(dead_code)]
-mod gen_topology;
-#[allow(dead_code)]
-mod gen_topology_files;
 mod node_run;
 mod utils;
