--- manifest.toml
@@ -1,18 +1,0 @@
-name = "smoke"
-
-[defaults]
-builder = "docker:generic"
-runner = "local:docker"
-
-[builders."docker:generic"]
-enabled = true
-
-[runners."local:docker"]
-enabled = true
-
-[[testcases]]
-name = "smoke"
-instances = { min = 2, max = 100, default = 2 }
-
-[testcases.params]
-gossipsub_history_length = { type = "int", desc = "Number of heartbeats to keep in the `memcache`", default = 5 }
