+++ utils/gen_topology/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "gen_topology"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+
+rand = "0.8.5"
+rand_chacha = "0.3.1"
+serde = { workspace = true, features = [ "derive" ]}
