+++ utils/gen_topology_files/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "gen_topology_files"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+gen_topology = { path = "../gen_topology" }
+clap = { version = "4.0.10", features = ["derive"] }
+serde_json.workspace = true
